Update Mielikki's Topiary
I would remove the lower level because the trapdoor is hidden.  Replace that room with an X or * and note it as a "point of interest".

I've added an extra - to each east-west path just to make it a little bigger and make the rows look more even.

I've also added a second version so you can see what it looks like with the color codes added - it LOOKS totally out of alignment, but all the &# will disappear when it's read, making it look like the one you created.  The garden rooms will be dark green, the paths between and the cobblestone path will be brown/orange/dark yellow, the gate and "Points of Interest" text will be bright yellow, and the gazebo will be bright white.  The point labels will be plain text and the zone change label will be turquoise.

For now I'm making all the fancy borders dark yellow/orange/brown and the zone name bright green.